MAINT: mypy and formatters
